Initial Awilix rewrite (#144)
* introduce Awilix IoC framework
* refactor code
* fix lint
