--- modules/GlobalEvents.js
@@ -1,4 +1,0 @@
-var events = require('events');
-
-var em = new events.EventEmitter();
-module.exports.globalEmitter = em;
