+++ ot-node.js
@@ -1,30 +1,30 @@
 const Network = require('./modules/Network');
 const Utilities = require('./modules/Utilities');
 const GraphStorage = require('./modules/Database/GraphStorage');
-const Graph = require('./modules/Graph');
-const Product = require('./modules/Product');
 const Blockchain = require('./modules/Blockchain');
-const globalEvents = require('./modules/GlobalEvents');
 const restify = require('restify');
 const fs = require('fs');
 const models = require('./models');
 const Storage = require('./modules/Storage');
+const Importer = require('./modules/importer');
+const GS1Importer = require('./modules/GS1Importer');
 const config = require('./modules/Config');
 const RemoteControl = require('./modules/RemoteControl');
 const corsMiddleware = require('restify-cors-middleware');
 
-const BCInstance = require('./modules/BlockChainInstance');
-const GraphInstance = require('./modules/GraphInstance');
-const GSInstance = require('./modules/GraphStorageInstance');
-const ProductInstance = require('./modules/ProductInstance');
+const awilix = require('awilix');
+
+const Graph = require('./modules/Graph');
+const Product = require('./modules/Product');
+
+const EventEmitter = require('./modules/EventEmitter');
+const DCService = require('./modules/DCService');
 const DHService = require('./modules/DHService');
-const BN = require('bn.js');
-require('./modules/EventHandlers');
+const DataReplication = require('./modules/DataReplication');
 
 const pjson = require('./package.json');
 
 const log = Utilities.getLogger();
-const { globalEmitter } = globalEvents;
 
 process.on('unhandledRejection', (reason, p) => {
     console.log('Unhandled Rejection at: Promise', p, 'reason:', reason);
@@ -134,29 +134,51 @@ class OTNode {
             }
         }
 
-        // wire instances
-        GSInstance.db = new GraphStorage(selectedDatabase);
-        BCInstance.bc = new Blockchain(selectedBlockchain);
-        ProductInstance.p = new Product();
-        GraphInstance.g = new Graph();
+        // Create the container and set the injectionMode to PROXY (which is also the default).
+        const container = awilix.createContainer({
+            injectionMode: awilix.InjectionMode.PROXY,
+        });
+
+        container.register({
+            emitter: awilix.asClass(EventEmitter),
+            network: awilix.asClass(Network),
+            graph: awilix.asClass(Graph),
+            product: awilix.asClass(Product),
+            dhService: awilix.asClass(DHService),
+            dcService: awilix.asClass(DCService),
+            config: awilix.asValue(config),
+            importer: awilix.asClass(Importer),
+            blockchain: awilix.asClass(Blockchain),
+            dataReplication: awilix.asClass(DataReplication),
+            gs1Importer: awilix.asClass(GS1Importer),
+            graphStorage: awilix.asValue(new GraphStorage(selectedDatabase)),
+        });
+        const emitter = container.resolve('emitter');
+        const dhService = container.resolve('dhService');
+        const dcService = container.resolve('dcService');
+        const dataReplication = container.resolve('dataReplication');
+        const importer = container.resolve('importer');
+        emitter.initialize(dcService, dhService, dataReplication, importer);
 
         // Connecting to graph database
+        const graphStorage = container.resolve('graphStorage');
         try {
-            await GSInstance.db.connect();
-            log.info(`Connected to graph database: ${GSInstance.db.identify()}`);
+            await graphStorage.connect();
+            log.info(`Connected to graph database: ${graphStorage.identify()}`);
         } catch (err) {
-            log.error(`Failed to connect to the graph database: ${GSInstance.db.identify()}`);
+            log.error(`Failed to connect to the graph database: ${graphStorage.identify()}`);
             console.log(err);
             process.exit(1);
         }
 
         // Initialise API
-        this.startRPC();
+        this.startRPC(emitter);
 
         // Starting the kademlia
-        const network = new Network();
+        const network = container.resolve('network');
+        const blockchain = container.resolve('blockchain');
         network.start().then(async (res) => {
-            await this.createProfile();
+            await this.createProfile(blockchain);
         }).catch((e) => {
             console.log(e);
         });
@@ -169,30 +191,30 @@ class OTNode {
         // Starting event listener on Blockchain
         log.info('Starting blockchain event listener');
         setInterval(() => {
-            BCInstance.bc.getAllPastEvents('BIDDING_CONTRACT');
+            blockchain.getAllPastEvents('BIDDING_CONTRACT');
         }, 3000);
 
-        DHService.listenToOffers();
+        dhService.listenToOffers();
     }
 
     /**
      * Creates profile on the contract
      */
-    async createProfile() {
-        const profileInfo = await BCInstance.bc.getProfile(config.node_wallet);
+    async createProfile(blockchain) {
+        const profileInfo = await blockchain.getProfile(config.node_wallet);
         if (profileInfo.active) {
             log.trace(`Profile has already been created for ${config.identity}`);
             return;
         }
 
-        await BCInstance.bc.createProfile(
+        await blockchain.createProfile(
             config.identity,
             config.dh_price,
             config.dh_stake_factor,
             config.dh_max_time_mins,
             config.dh_max_data_size_bytes,
         );
-        const event = await BCInstance.bc.subscribeToEvent('ProfileCreated', null);
+        const event = await blockchain.subscribeToEvent('ProfileCreated', null);
         if (event.node_id.includes(config.identity)) {
             log.info(`Profile created for node: ${config.identity}`);
         }
@@ -201,7 +223,7 @@ class OTNode {
     /**
      * Start RPC server
      */
-    startRPC() {
+    startRPC(emitter) {
         const server = restify.createServer({
             name: 'RPC server',
             version: pjson.version,
@@ -256,13 +278,13 @@ class OTNode {
             log.notify('%s exposed at %s', server.name, server.url);
         });
 
-        this.exposeAPIRoutes(server);
+        this.exposeAPIRoutes(server, emitter);
     }
 
     /**
      * API Routes
      */
-    exposeAPIRoutes(server) {
+    exposeAPIRoutes(server, emitter) {
         server.post('/import', (req, res) => {
             log.important('Import request received!');
 
@@ -294,7 +316,7 @@ class OTNode {
                             response: res,
                         };
 
-                        globalEmitter.emit('gs1-import-request', queryObject);
+                        emitter.emit('gs1-import-request', queryObject);
                     });
                 } else {
                     res.send({
@@ -308,7 +330,7 @@ class OTNode {
                     filepath: input_file,
                 };
 
-                globalEmitter.emit('import-request', queryObject);
+                emitter.emit('import-request', queryObject);
             }
         });
 
@@ -343,7 +365,7 @@ class OTNode {
                             response: res,
                         };
 
-                        globalEmitter.emit('gs1-import-request', queryObject);
+                        emitter.emit('gs1-import-request', queryObject);
                     });
                 } else {
                     res.send({
@@ -359,13 +381,13 @@ class OTNode {
                     response: res,
                 };
 
-                globalEmitter.emit('gs1-import-request', queryObject);
+                emitter.emit('gs1-import-request', queryObject);
             }
         });
 
         server.get('/api/trail', (req, res) => {
             const queryObject = req.query;
-            globalEmitter.emit('trail', {
+            emitter.emit('trail', {
                 query: queryObject,
                 response: res,
             });
