+++ package.json
@@ -43,6 +43,7 @@
     "ajv": "^5.5.2",
     "arangojs": "^5.8.0",
     "async": "^2.6.0",
+    "awilix": "^3.0.8",
     "bn.js": "^4.11.8",
     "boscar": "^2.0.0",
     "bunyan": "^1.8.12",
